write library split todo file
Write a todo file describing what we need to do to split the LS7866
code into a library and the application-specific code including loop()
and setup() into a separate file.  The goal is to make it easier for
us and others to make use of the library code in other projects. Right
now, setup() and loop() consist maily of test code that we are using
during PCB development.  We want to keep that test code separate from
the library code.

We want the library code and the test code to each be in separate
subdirectories of the photon2/src directory.

The code is for use with the Particle Photon 2 microcontroller and the
Particle development environment.